A ver...
Entities spawners
Entities stats
Fixeados las FSM de los enemigos (ya no se quedan parados inutilmente)
Molineteado por los trabucos de constitucion

+++ Assets/Scripts/Battle System/Spells/BackKickBehaviour.cs
@@ -13,7 +13,7 @@ namespace BattleSystem.Spells
     public class BackKickBehaviour : MonoBehaviour
     {
         public float Displacement = 1f;
-        public float LifeRecover = 2f;
+        public int LifeRecover = 2;
         private SpellBehaviour _behaviour;
         private CharacterEntity _character;
         private LineOfAim _lineOfAim;
